requirements.txt
asgiref==3.7.2
astroid==2.11.7
autopep8==2.0.2
beautifulsoup4==4.11.1
black==23.1.0
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
colorgram.py==1.2.0
contourpy==1.0.7
crispy-bootstrap4==2022.1
cryptography==39.0.1
cycler==0.11.0
dill==0.3.5.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.2.5
django-ckeditor==6.5.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==1.2.2
django-pandas==0.6.6
django-storages==1.13.1
django-unused-media==0.2.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.7.1
fingerprint==0.1.6
fonttools==4.38.0
html5lib==1.1
idna==3.3
imutils==0.5.4
install==1.3.5
isort==5.10.1
keras==2.13.1
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
matplotlib==3.7.1
mccabe==0.7.0
msedge-selenium-tools==3.141.4
mypy-extensions==1.0.0
newsapi==0.1.1
newsapi-python==0.2.6
numpy==1.22.3

openpyxl==3.1.1
packaging==23.0
pandas==1.4.2
pathspec==0.11.0
Pillow==9.1.0
platformdirs==2.5.2
plotly==5.13.1
praw==7.6.0
prawcore==2.3.0
psycopg2==2.9.3
pycodestyle==2.10.0
pycparser==2.21
pylint==2.14.5
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-barcode==0.14.0
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
pytz==2020.1

qrcode==7.3.1
requests==2.27.1
selenium==3.141.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.3.1
starkbank-ecdsa==2.0.3
tenacity==8.2.2
tk==0.1.0
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.7.1
tzdata==2022.1
update-checker==0.18.0
urllib3==1.25.9
virtualenv==20.14.1
virtualenvwrapper-win==1.2.7
webencodings==0.5.1
websocket-client==1.4.1
whitenoise==5.1.0
windows-curses==2.3.0
wrapt==1.14.1
